Correct the scavenging code.
Change the card table to cover all fields, not just the mutable ones.
This is needed since we are not doing eager promotion, and as a result
evacuation might fail.  If this happens, then we set the bit in the card
table, to remember to scavange the field next time around.